Refactor MySQLCommandPacketDecoder
