+++ shardingsphere-scaling/shardingsphere-scaling-dialect/shardingsphere-scaling-mysql/src/test/java/org/apache/shardingsphere/scaling/mysql/client/netty/MySQLCommandPacketDecoderTest.java
@@ -18,15 +18,9 @@
 package org.apache.shardingsphere.scaling.mysql.client.netty;
 
 import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufUtil;
-import io.netty.buffer.Unpooled;
-import org.apache.shardingsphere.db.protocol.mysql.constant.MySQLServerInfo;
-import org.apache.shardingsphere.db.protocol.mysql.constant.MySQLStatusFlag;
 import org.apache.shardingsphere.db.protocol.mysql.packet.generic.MySQLEofPacket;
 import org.apache.shardingsphere.db.protocol.mysql.packet.generic.MySQLErrPacket;
 import org.apache.shardingsphere.db.protocol.mysql.packet.generic.MySQLOKPacket;
-import org.apache.shardingsphere.db.protocol.mysql.packet.handshake.MySQLHandshakePacket;
-import org.apache.shardingsphere.scaling.core.util.ReflectionUtil;
 import org.apache.shardingsphere.scaling.mysql.client.InternalResultSet;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -47,56 +41,10 @@ public final class MySQLCommandPacketDecoderTest {
     @Mock
     private ByteBuf byteBuf;
     
-    @Test(expected = IllegalArgumentException.class)
-    public void assertDecodeUnsupportedProtocolVersion() {
-        MySQLCommandPacketDecoder commandPacketDecoder = new MySQLCommandPacketDecoder();
-        commandPacketDecoder.decode(null, byteBuf, null);
-    }
-    
-    @Test(expected = UnsupportedOperationException.class)
-    public void assertDecodeUnsupportedAuthenticationMethod() {
-        when(byteBuf.readUnsignedByte()).thenReturn((short) 0, (short) MySQLServerInfo.PROTOCOL_VERSION);
-        when(byteBuf.readUnsignedShortLE()).thenReturn(MySQLStatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue());
-        MySQLCommandPacketDecoder commandPacketDecoder = new MySQLCommandPacketDecoder();
-        commandPacketDecoder.decode(null, byteBuf, null);
-    }
-    
-    @Test
-    public void assertDecodeHandshakePacket() {
-        MySQLCommandPacketDecoder commandPacketDecoder = new MySQLCommandPacketDecoder();
-        List<Object> actual = new LinkedList<>();
-        commandPacketDecoder.decode(null, mockHandshakePacket(), actual);
-        assertHandshakePacket(actual);
-    }
-    
-    private ByteBuf mockHandshakePacket() {
-        String handshakePacket = "000a352e372e32312d6c6f6700090000004a592a1f725a0d0900fff7210200ff8115000000000000000000001a437b30323a4d2b514b5870006d"
-            + "7973716c5f6e61746976655f70617373776f72640000000002000000";
-        byte[] handshakePacketBytes = ByteBufUtil.decodeHexDump(handshakePacket);
-        ByteBuf result = Unpooled.buffer(handshakePacketBytes.length);
-        result.writeBytes(handshakePacketBytes);
-        return result;
-    }
-    
-    private void assertHandshakePacket(final List<Object> actual) {
-        assertThat(actual.size(), is(1));
-        assertThat(actual.get(0), instanceOf(MySQLHandshakePacket.class));
-        MySQLHandshakePacket actualPacket = (MySQLHandshakePacket) actual.get(0);
-        assertThat(actualPacket.getProtocolVersion(), is(0x0a));
-        assertThat(actualPacket.getServerVersion(), is("5.7.21-log"));
-        assertThat(actualPacket.getConnectionId(), is(9));
-        assertThat(actualPacket.getCharacterSet(), is(33));
-        assertThat(actualPacket.getStatusFlag().getValue(), is(2));
-        assertThat(actualPacket.getCapabilityFlagsLower(), is(63487));
-        assertThat(actualPacket.getCapabilityFlagsUpper(), is(33279));
-        assertThat(actualPacket.getAuthPluginName(), is("mysql_native_password"));
-    }
-    
     @Test
     public void assertDecodeOkPacket() throws NoSuchFieldException, IllegalAccessException {
         MySQLCommandPacketDecoder commandPacketDecoder = new MySQLCommandPacketDecoder();
         List<Object> actual = new LinkedList<>();
-        ReflectionUtil.setFieldValue(commandPacketDecoder, "authenticated", true);
         commandPacketDecoder.decode(null, mockOkPacket(), actual);
         assertPacketByType(actual, MySQLOKPacket.class);
     }
@@ -111,7 +59,6 @@ public final class MySQLCommandPacketDecoderTest {
     public void assertDecodeErrPacket() throws NoSuchFieldException, IllegalAccessException {
         MySQLCommandPacketDecoder commandPacketDecoder = new MySQLCommandPacketDecoder();
         List<Object> actual = new LinkedList<>();
-        ReflectionUtil.setFieldValue(commandPacketDecoder, "authenticated", true);
         commandPacketDecoder.decode(null, mockErrPacket(), actual);
         assertPacketByType(actual, MySQLErrPacket.class);
     }
@@ -126,7 +73,6 @@ public final class MySQLCommandPacketDecoderTest {
     public void assertDecodeQueryCommPacket() throws NoSuchFieldException, IllegalAccessException {
         MySQLCommandPacketDecoder commandPacketDecoder = new MySQLCommandPacketDecoder();
         List<Object> actual = new LinkedList<>();
-        ReflectionUtil.setFieldValue(commandPacketDecoder, "authenticated", true);
         commandPacketDecoder.decode(null, mockEmptyResultSetPacket(), actual);
         commandPacketDecoder.decode(null, mockFieldDefinition41Packet(), actual);
         commandPacketDecoder.decode(null, mockEofPacket(), actual);
